+++ tests/db/test_cursor.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
+from __future__ import unicode_literals
 
 import unittest
+import warnings
 from collections import namedtuple
 
 from requests.models import Response
@@ -19,14 +21,13 @@ class CursorTestSuite(unittest.TestCase):
     def test_execute(self, requests_post_mock):
         response = Response()
         response.status_code = 200
-        response.raw = BytesIO(
-            b'[{"name": "alice"}, {"name": "bob"}, {"name": "charlie"}]'
-        )
+        response.raw = BytesIO(b'["name"]\n["alice"]\n["bob"]\n["charlie"]\n\n')
         requests_post_mock.return_value = response
         Row = namedtuple("Row", ["name"])
 
         cursor = Cursor("http://example.com/")
         cursor.execute("SELECT * FROM table")
+        self.assertEqual(cursor.description, [("name", None)])
         result = cursor.fetchall()
         expected = [Row(name="alice"), Row(name="bob"), Row(name="charlie")]
         self.assertEqual(result, expected)
@@ -35,14 +36,33 @@ class CursorTestSuite(unittest.TestCase):
     def test_execute_empty_result(self, requests_post_mock):
         response = Response()
         response.status_code = 200
-        response.raw = BytesIO(b"[]")
+        response.raw = BytesIO(b'["name"]\n\n')
         requests_post_mock.return_value = response
 
         cursor = Cursor("http://example.com/")
         cursor.execute("SELECT * FROM table")
+        self.assertEqual(cursor.description, [("name", None)])
         result = cursor.fetchall()
         expected = []
         self.assertEqual(result, expected)
+
+    @patch("requests.post")
+    def test_truncated_response(self, requests_post_mock):
+        response = Response()
+        response.status_code = 200
+        response.raw = BytesIO(b'["name"]\n["alice"]\n')
+        requests_post_mock.return_value = response
+
+        cursor = Cursor("http://example.com/")
+        cursor.execute("SELECT * FROM table")
+        self.assertEqual(cursor.description, [("name", None)])
+
+        with self.assertRaises(ValueError) as cm:
+            cursor.fetchall()
+
+        self.assertEqual(
+            cm.exception.args[0], "Truncated response. Trailer line not found."
+        )
 
     @patch("requests.post")
     def test_context(self, requests_post_mock):
@@ -63,7 +83,12 @@ class CursorTestSuite(unittest.TestCase):
             auth=None,
             stream=True,
             headers={"Content-Type": "application/json"},
-            json={"query": query, "context": context, "header": False},
+            json={
+                "query": query,
+                "context": context,
+                "header": True,
+                "resultFormat": "arrayLines",
+            },
             verify=True,
             cert=None,
             proxies=None,
@@ -73,27 +98,31 @@ class CursorTestSuite(unittest.TestCase):
     def test_header_false(self, requests_post_mock):
         response = Response()
         response.status_code = 200
-        response.raw = BytesIO(b'[{"name": "alice"}]')
+        response.raw = BytesIO(b'["name"]\n["alice"]\n\n')
         requests_post_mock.return_value = response
         Row = namedtuple("Row", ["name"])
 
-        url = "http://example.com/"
-        query = "SELECT * FROM table"
+        with warnings.catch_warnings(record=True) as warning_list:
+            warnings.simplefilter("always")
 
-        cursor = Cursor(url, header=False)
-        cursor.execute(query)
+            cursor = Cursor("http://example.com/", header=False)
+
+        self.assertIn(
+            "Disabling the `header` parameter is not supported in this version of the lib."  # noqa: E501
+            " The value will be ignored and we will force `header=True`.",
+            str(warning_list[-1].message),
+        )
+
+        cursor.execute("SELECT * FROM table")
         result = cursor.fetchall()
         self.assertEqual(result, [Row(name="alice")])
-
-        self.assertEqual(
-            cursor.description, [("name", 1, None, None, None, None, True)]
-        )
+        self.assertEqual(cursor.description, [("name", None)])
 
     @patch("requests.post")
     def test_header_true(self, requests_post_mock):
         response = Response()
         response.status_code = 200
-        response.raw = BytesIO(b'[{"name": null}, {"name": "alice"}]')
+        response.raw = BytesIO(b'["name"]\n["alice"]\n\n')
         requests_post_mock.return_value = response
         Row = namedtuple("Row", ["name"])
 
@@ -110,14 +139,14 @@ class CursorTestSuite(unittest.TestCase):
     def test_names_with_underscores(self, requests_post_mock):
         response = Response()
         response.status_code = 200
-        response.raw = BytesIO(b'[{"_name": null}, {"_name": "alice"}]')
+        response.raw = BytesIO(b'["_name"]\n["alice"]\n\n')
         requests_post_mock.return_value = response
         Row = namedtuple("Row", ["_name"], rename=True)
 
         url = "http://example.com/"
         query = "SELECT * FROM table"
 
-        cursor = Cursor(url, header=True)
+        cursor = Cursor(url)
         cursor.execute(query)
         result = cursor.fetchall()
         self.assertEqual(result, [Row(_0="alice")])